requirements.txt
aniso8601==8.0.0
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
Click==7.0
decorator==4.4.1
Flask==1.1.1
flask-restplus==0.13.0
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.4.1
greenlet==0.4.15
importlib-metadata==0.23
ipython==7.9.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
msgpack==0.6.2
neovim==0.3.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
pynvim==0.3.2
pyrsistent==0.15.5
pytz==2019.3
six==1.12.0
SQLAlchemy==1.3.10
traitlets==4.3.3
uWSGI==2.0.18
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
Flask-APScheduler==1.11.0